+++ nest_lesson_3/src/modules/testingModule/testing.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { TestingController } from './controller/testing.controller';
+import { BlogsModule } from '../blogsModule/blogs.module';
+import { UsersModule } from '../usersModule/users.module';
+
+@Module({
+  imports: [BlogsModule, UsersModule],
+  controllers: [TestingController],
+  providers: [],
+  exports: [],
+})
+export class TestingModule {}
